requirements.txt
minimalmodbus==1.0.2
pyserial==3.4
python-socketio==4.6.0
requests